requirements.txt
appnope==0.1.0
backports.ssl-match-hostname==3.4.0.2
certifi==2015.9.6.2
cycler==0.9.0
decorator==4.0.4
-e git+https://github.com/jwass/directions.py@e30a9bfda1faf367def49a919444727bda47eb50#egg=directions.py-master
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-annoying==0.8.4
django-rest-swagger==0.3.4
djangocms-admin-style==0.2.8
djangorestframework==3.3.0
djangorestframework-jwt==1.7.2
functools32==3.2.3.post2
gnureadline==6.3.3
gunicorn==19.3.0
ipykernel==4.1.1
ipython==4.0.0
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.1.1
jupyter-core==4.0.6
MarkupSafe==0.23
matplotlib==1.5.0
mistune==0.7.1
nbconvert==4.1.0
nbformat==4.0.1
notebook==4.0.6
numpy==1.10.1
oauthlib==1.0.3
overpy==0.3.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
-e git+https://github.com/Mapkin/polycomp/@28a8d9a5ece1e58308dffbd087e1a2be3ddf8c49#egg=polycomp-master
ptyprocess==0.5
Pygments==2.0.2
PyJWT==1.4.0
pyparsing==2.0.5
pystache==0.5.4
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
pyzmq==14.7.0
requests==2.8.1
requests-oauthlib==0.5.0
scipy==0.16.1
simplegeneric==0.8.1
six==1.10.0
static==1.1.1
static3==0.6.1
terminado==0.5
tornado==4.2.1
traitlets==4.0.0
wheel==0.24.0
